prep for curve support and fix bad deploys dataentry
